api/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers>=4.4.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
python-dotenv
Pillow==10.4.0
django-filter==24.2
dj-database-url>=2.2.0
psycopg[binary]
whitenoise==6.6.0
gunicorn>=21.2.0
weasyprint>=60.0
qrbill>=1.1.0
cairosvg>=2.7.0
swiss-qr-bill>=0.0.31
paddleocr>=2.7.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.0
